requirements.txt
telethon==1.40.0 
python-socketio==5.13.0 
starlette==0.46.2 
uvicorn==0.34.2
python-dotenv